pta: qcom: pas: refactor into a descriptor/ops abstraction
Move all Kodiak-specific logic and the PTA command handlers into
platform/kodiak/, and model each subsystem with a descriptor/ops
abstraction: every platform exposes a table via
qcom_pas_platform_subsys() that the generic pas_core.c drives.
Pure structural refactor.
