+++ core/arch/arm/plat-qcom/hoya/kodiak/target_config.h
@@ -39,4 +39,9 @@
 #define IRIS_BASE			UL(0x0aa00000)
 #define IRIS_SIZE			ULL(0x00200000)
 
+#define PAS_ID_QDSP6			1
+#define PAS_ID_WPSS			6
+#define PAS_ID_VENUS			9
+#define PAS_ID_TURING			18
+
 #endif /* TARGET_CONFIG_H */
