+++ core/pta/qcom/pas/platform/kodiak/cdsp.h
@@ -1,0 +1,13 @@
+/* SPDX-License-Identifier: BSD-2-Clause */
+/*
+ * Copyright (c) 2026, Qualcomm Technologies, Inc. and/or its subsidiaries.
+ */
+
+#ifndef _CDSP_H_
+#define _CDSP_H_
+
+#include "pas_subsys.h"
+
+extern const struct qcom_pas_ops cdsp_ops;
+
+#endif /* _CDSP_H_ */
