+++ core/pta/qcom/pas/platform/kodiak/lpass.h
@@ -1,0 +1,13 @@
+/* SPDX-License-Identifier: BSD-2-Clause */
+/*
+ * Copyright (c) 2026, Qualcomm Technologies, Inc. and/or its subsidiaries.
+ */
+
+#ifndef _LPASS_H_
+#define _LPASS_H_
+
+#include "pas_subsys.h"
+
+extern const struct qcom_pas_ops lpass_ops;
+
+#endif /* _LPASS_H_ */
