+++ core/pta/qcom/pas/platform/kodiak/sub.mk
@@ -1,0 +1,3 @@
+srcs-y += subsys.c cdsp.c dsp_fw.c lpass.c wpss.c venus.c
+incdirs-y += .
+incdirs-y += ../
