+++ core/pta/qcom/pas/platform/kodiak/venus.h
@@ -3,10 +3,14 @@
  * Copyright (c) 2026, Qualcomm Technologies, Inc. and/or its subsidiaries.
  */
 
-#ifndef _VIDEO_H_
-#define _VIDEO_H_
+#ifndef _VENUS_H_
+#define _VENUS_H_
+
+#include "pas_subsys.h"
 
 #define IRIS_WRAPPER_TOP_TZ_REG_BASE	0x000c0000
 #define IRIS_WRAPPER_TOP_REG_BASE	0x000b0000
 
-#endif
+extern const struct qcom_pas_ops venus_ops;
+
+#endif /* _VENUS_H_ */
