+++ core/pta/qcom/pas/platform/kodiak/wpss.h
@@ -1,0 +1,13 @@
+/* SPDX-License-Identifier: BSD-2-Clause */
+/*
+ * Copyright (c) 2026, Qualcomm Technologies, Inc. and/or its subsidiaries.
+ */
+
+#ifndef _WPSS_H_
+#define _WPSS_H_
+
+#include "pas_subsys.h"
+
+extern const struct qcom_pas_ops wpss_ops;
+
+#endif /* _WPSS_H_ */
